Split software into separate server and FTP client parts; add systemd services
Software now runs as two separate systemd services: the server, which operates the magnetometer and exposes
a RESTful web interface; and the FTP client which synchronises the data with an external FTP server.

Running setup.py now copies the systemd service configurations to the necessary directory, if run on a
supported platform.